feat(backend): buf change, login service
Remove legacy api/old protobuf definitions, add new xyz.city_ideas.v1 protos and googleapis submodule, and update buf configuration and Go generation. .gitmodules: remove grpc-web submodule; api/buf.gen.yaml and api/buf.yaml: change plugin output to backend (module github.com/aesterial/cityideas/backend), add default_api_level, adjust require_unimplemented_servers and exclude/ignore third_party. Regenerated/added Go protobuf outputs under backend/internal/api/v1/* and moved/refactored server interceptor/handler files (renames), added login_handler.go, and applied related backend code updates (main, services, config, repositories, migrations) to match the new proto layout. This consolidates proto layout, cleans up old definitions, and updates generation paths/options for the codebase.

+++ backend/migrations/scheme/scheme.sql
@@ -83,7 +83,7 @@ create table if not exists ranks
     id          uuid        not null default gen_random_uuid(),
     name        text                   not null,
     description text                   not null default '',
-    color       int                    not null default 0,
+    color       bigint                 not null default 0,
     weight      int                    not null default 0,
     permissions jsonb                  not null,
     added_at    timestamptz not null default now()
